requirements.txt
alembic
fastapi==0.79.0
sqlalchemy
GeoAlchemy2
psycopg2-binary
uvicorn
python-jose
passlib
starlette==0.19.1
